+++ src/clj/cascalog/playground.clj
@@ -1,18 +1,3 @@
-;;    Copyright 2011 Nathan Marz
-;; 
-;;    This program is free software: you can redistribute it and/or modify
-;;    it under the terms of the GNU General Public License as published by
-;;    the Free Software Foundation, either version 3 of the License, or
-;;    (at your option) any later version.
-;; 
-;;    This program is distributed in the hope that it will be useful,
-;;    but WITHOUT ANY WARRANTY; without even the implied warranty of
-;;    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-;;    GNU General Public License for more details.
-;; 
-;;    You should have received a copy of the GNU General Public License
-;;    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
 (ns cascalog.playground
   (:require [cascalog.conf :as conf])
   (:import [java.io PrintStream]
@@ -34,10 +19,10 @@
   (System/setOut (PrintStream. (WriterOutputStream. *out*))))
 
 (defn bootstrap-jcascalog []
-  (import '[cascalog StdoutTap])
-  (import '[jcascalog Api Fields Predicate PredicateMacro Subquery
-            Api$FirstNArgs Option])
-  (import '[jcascalog.op Sum Count Div Plus Minus Multiply Avg
+  (import '[com.twitter.maple.tap StdoutTap]
+          '[jcascalog Api Fields Predicate PredicateMacro Subquery
+            Api$FirstNArgs Option]
+          '[jcascalog.op Sum Count Div Plus Minus Multiply Avg
             Max Min Limit LimitRank ReParse DistinctCount]))
 
 (def person