+++ src/clj/cascalog/rules.clj
@@ -1,18 +1,3 @@
-;;    Copyright 2010 Nathan Marz
-;; 
-;;    This program is free software: you can redistribute it and/or modify
-;;    it under the terms of the GNU General Public License as published by
-;;    the Free Software Foundation, either version 3 of the License, or
-;;    (at your option) any later version.
-;; 
-;;    This program is distributed in the hope that it will be useful,
-;;    but WITHOUT ANY WARRANTY; without even the implied warranty of
-;;    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-;;    GNU General Public License for more details.
-;; 
-;;    You should have received a copy of the GNU General Public License
-;;    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
 (ns cascalog.rules
   (:use [cascalog.debug :only (debug-print)]
         [clojure.set :only (intersection union difference subset?)]
@@ -30,9 +15,10 @@
            [cascading.tuple Fields Tuple TupleEntry]
            [cascading.flow Flow FlowConnector]
            [cascading.pipe Pipe]
-           [cascading.pipe.cogroup CascalogJoiner CascalogJoiner$JoinType]
+           [cascading.flow.hadoop HadoopFlowProcess]
+           [cascading.pipe.joiner CascalogJoiner CascalogJoiner$JoinType]
            [cascalog CombinerSpec ClojureCombiner ClojureCombinedAggregator Util
-                     ClojureParallelAgg]
+            ClojureParallelAgg]
            [org.apache.hadoop.mapred JobConf]
            [jcascalog Predicate Subquery PredicateMacro ClojureOp]
            [java.util ArrayList]))
@@ -93,7 +79,8 @@
 (defn- fixed-point-operations
   "Adds operations to tail until can't anymore. Returns new tail"
   [tail]
-  (if-let [op (s/find-first (partial op-allowed? tail) (:operations tail))]
+  (if-let [op (s/find-first (partial op-allowed? tail)
+                            (:operations tail))]
     (recur (add-op tail op))
     tail))
 
@@ -417,7 +404,7 @@
               :outfields project-fields}))))
 
 (def DEFAULT-OPTIONS
-  {:distinct true
+  {:distinct false
    :sort nil
    :reverse false
    :trap nil})
@@ -556,8 +543,7 @@
                                                  (mapcat split-outvar-constants)
                                                  (map rewrite-predicate)
                                                  (mapcat split-outvar-constants)
-                                                 (uniquify-query-vars out-vars)
-                                                 )
+                                                 (uniquify-query-vars out-vars))
         [raw-opts raw-predicates] (s/separate #(keyword? (first %)) raw-predicates)
         options                   (mk-options (map p/mk-option-predicate raw-opts))
         [gens ops aggs]           (->> raw-predicates
@@ -666,8 +652,9 @@
 (defn mk-raw-predicate [[op-sym & vars]]
   [op-sym (v/vars->str vars)])
 
-(defn- pluck-tuple [tap]
-  (with-open [it (.openForRead tap (hadoop/job-conf (conf/project-conf)))]
+(defn pluck-tuple [tap]
+  (with-open [it (-> (HadoopFlowProcess. (hadoop/job-conf (conf/project-conf)))
+                     (.openTapForRead tap))]
     (if-let [iter (iterator-seq it)]
       (-> iter first .getTuple Tuple. Util/coerceFromTuple vec)
       (throw-illegal "Cascading tap is empty -- tap must contain tuples."))))
@@ -683,21 +670,25 @@ cascading tap, returns a new generator with field-names."
         (or (instance? Tap g)
             (vector? g)
             (list? g))
-        (let [pluck (if (instance? Tap g) pluck-tuple first)
-              vars  (v/gen-nullable-vars (count (pluck g)))]
-          (->> [[g :>> vars] [:distinct false]]
-               (map mk-raw-predicate)
-               (build-rule vars)))
+        (let [pluck (if (instance? Tap g) pluck-tuple, first)
+              size  (count (pluck g))
+              vars  (v/gen-nullable-vars size)]
+          (if (zero? size)
+            (throw-illegal
+             "Data structure is empty -- memory sources must contain tuples.")
+            (->> [[g :>> vars] [:distinct false]]
+                 (map mk-raw-predicate)
+                 (build-rule vars))))
         :else g))
 
+;; TODO: Why does this not use gen?
 (defn connect-to-sink [gen sink]
   ((w/pipe-rename (u/uuid)) (:pipe gen)))
 
 (defn normalize-gen [gen]
   (if (instance? Subquery gen)
     (.getCompiledSubquery gen)
-    gen
-    ))
+    gen))
 
 (defn combine* [gens distinct?]
   ;; it would be nice if cascalog supported Fields/UNKNOWN as output of generator
@@ -740,9 +731,15 @@ cascading tap, returns a new generator with field-names."
     ["" args]))
 
 (defn get-sink-tuples [^Tap sink]
-  (if (map? sink)
-    (get-sink-tuples (:sink sink))
-    (with-open [it (.openForRead sink (hadoop/job-conf (conf/project-conf)))]
-      (doall
-       (for [^TupleEntry t (iterator-seq it)]
-         (vec (Util/coerceFromTuple (Tuple. (.getTuple t)))))))))
+  (let [conf (hadoop/job-conf (conf/project-conf))]
+    (cond (map? sink)
+          (get-sink-tuples (:sink sink))
+
+          (not (.resourceExists sink conf))
+          []
+
+          :else (with-open [it (-> (HadoopFlowProcess. conf)
+                                   (.openTapForRead sink))]
+                  (doall
+                   (for [^TupleEntry t (iterator-seq it)]
+                     (into [] (Tuple. (.getTuple t)))))))))