+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+### 1.9.0
+
+* Cascalog now uses the Apache 2.0 license (formerly used GNU)
+* Support for Cascading 2.0 (Hadoop planner only, for now)
+* Cascalog traps no longer trap the entire failing tuple; just the specific fields that caused the failure.
+* Sinkmode options are now :keep, :update and :replace, vs the old :keep, :append and :replace.
+* Extensive documentation of functions in cascalog.ops
+* Cascalog now pulls in Clojure 1.4.0 by default (with testing against 1.2 and 1.3)
+* `cascalog.StdoutTap` moved to `com.twitter.maple.StdoutTap`
+* `:distinct` defaults to false instead of true (!!)
+* Bug fix: Cascalog does a better job mapping functions to vars, doesn't get confused by "recent vars" (*1, *2, etc)
+
 ## 1.8.7
 
 * Exposed a pure-Java, fully-interoperable API to Cascalog called JCascalog